gmbp_quant/dpl/data_alert_bot/requirement.txt
aiohttp==3.7.4.post0
astcheck==0.2.5
astsearch==0.2.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cmdstanpy==0.9.68
convertdate==2.3.2
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
ephem==4.0.0.2
ftfy==6.0.1
gmbp-common==0.0.6
greenlet==1.0.0
hijri-converter==2.1.2
holidays==0.11.1
httplib2==0.19.1
idna==2.10
ipython-genutils==0.2.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nbparameterise==0.5
nest-asyncio==1.5.1
ntlm-auth==1.5.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
Pillow==8.2.0
plotly==4.14.3
prophet==1.0.1
pycparser==2.20
Pygments==2.9.0
PyMeeus==0.5.11
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
pystan==2.19.1.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pywin32==301
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-ntlm==1.1.0
retrying==1.3.3
schedule==1.1.0
scipy==1.6.2
secutils==0.0.3
setuptools-git==1.2
six==1.15.0
slack==0.0.2
slack-bolt==1.6.1
slack-sdk==3.6.0
slackclient==2.9.3
soupsieve==2.2.1
SQLAlchemy==1.4.8
tabulate==0.8.9
testpath==0.5.0
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
trading-calendars==2.1.1
traitlets==5.0.5
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.6.3
